Move imagemin to image-minify Move minify to babel-minify Make minify an aggregate
