+++ packages/babel-minify/.npmignore
