+++ packages/babel-minify/index.js
@@ -1,0 +1,5 @@
+const BabelMinify = require('babel-minify-webpack-plugin');
+
+module.exports = ({ config }, options = {}) => config
+  .plugin('babel-minify')
+  .use(BabelMinify, [options.minify, options.plugin]);
