+++ packages/image-minify/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@neutrinojs/imagemin",
+  "name": "@neutrinojs/image-minify",
   "version": "7.3.2",
   "description": "Neutrino middleware for optimizing images with imagemin",
   "main": "index.js",
@@ -11,7 +11,7 @@
   ],
   "author": "Tim Kelty <timkelty@gmail.com>",
   "license": "MPL-2.0",
-  "repository": "https://github.com/mozilla-neutrino/neutrino-dev/tree/master/packages/imagemin",
+  "repository": "https://github.com/mozilla-neutrino/neutrino-dev/tree/master/packages/image-minify",
   "homepage": "https://neutrino.js.org",
   "bugs": "https://github.com/mozilla-neutrino/neutrino-dev/issues",
   "dependencies": {
