+++ packages/image-minify/test/middleware_test.js
